+++ tests/test_number.py
@@ -42,7 +42,7 @@ async def test_hmfloat(factory: helper.Factory) -> None:
         value=0.3,
     )
     assert efloat.value == 0.3
-    central.event(const.INTERFACE_ID, "VCU0000011:3", "LEVEL", 0.5)
+    await central.event(const.INTERFACE_ID, "VCU0000011:3", "LEVEL", 0.5)
     assert efloat.value == 0.5
     # do not write. value above max
     await efloat.send_value(45.0)
@@ -116,7 +116,7 @@ async def test_hminteger(factory: helper.Factory) -> None:
     )
     assert einteger.value == 1
 
-    central.event(const.INTERFACE_ID, "VCU4984404:1", "SET_POINT_MODE", 2)
+    await central.event(const.INTERFACE_ID, "VCU4984404:1", "SET_POINT_MODE", 2)
     assert einteger.value == 2
     await einteger.send_value(6)
     assert mock_client.method_calls[-1] != call.set_value(